backend/tournament/requirements.txt
django>=5.0.0,<5.1.0
djangorestframework
pyyaml
requests
django-cors-headers
psycopg2-binary
python-dotenv
dj_database_url
Pillow
gunicorn
django-cors-headers
django-sslserver
whiteNoise
djangorestframework-simplejwt
web3
py-solc-x